test: convert tests from easymock to mockito for Storage.create() methods
